FS: F2FS: shrinker upstream commit2 fix warning build
fs/f2fs/super.c:73:19: предупреждение: initialization from incompatible pointer type
  .scan_objects  = f2fs_shrink_scan,
                   ^
fs/f2fs/super.c:73:19: предупреждение: (где-то рядом с инициализацией для «f2fs_shrinker_info.scan_objects»)

+++ fs/f2fs/shrinker.c
@@ -39,7 +39,7 @@ static unsigned long __count_extent_cache(struct f2fs_sb_info *sbi)
 				atomic_read(&sbi->total_ext_node);
 }
 
-int f2fs_shrink_count(struct shrinker *shrink,
+unsigned long f2fs_shrink_count(struct shrinker *shrink,
 				struct shrink_control *sc)
 {
 	struct f2fs_sb_info *sbi;
@@ -75,7 +75,7 @@ int f2fs_shrink_count(struct shrinker *shrink,
 	return count;
 }
 
-int f2fs_shrink_scan(struct shrinker *shrink,
+unsigned long f2fs_shrink_scan(struct shrinker *shrink,
 				struct shrink_control *sc)
 {
 	unsigned long nr = sc->nr_to_scan;
